requirements.txt
streamlit>=1.30.0
yfinance>=0.2.30
pandas>=2.0.0
requests>=2.31.0
pytz>=2024.1
python-dotenv>=1.0.0
pycryptodome>=3.20.0
